geotiff: dtype-aware GB hint + decimal GB units (#2553)
Addresses the /review-pr feedback on the initial commit:

* Switch _gb_hint to decimal GB (10**9) so the printed number matches
  the "~4 GB" claim in MAX_PIXELS_DEFAULT's docstring (and what df /
  du / S3 report).

* Plumb the decoded dtype through _check_dimensions and _gb_hint so
  the byte estimate uses dtype.itemsize. The hint now scales: f32 -> 4
  bytes/pixel, f64 -> 8, u8 -> 1, etc. Real call sites in _decode.py
  and _vrt.py pass the dtype; dtype= remains optional with the
  float32 fallback for callers without a resolved dtype.

* Add monkeypatched test isolating the VRT per-source error path,
  closing the nit about that string being covered only via the
  dimension guard's overlap.

* New tests assert dtype scaling at the helper, layout-guard, and
  VRT-source-guard levels.

+++ xrspatial/geotiff/_decode.py
@@ -383,7 +383,7 @@ def _read_strips(data: bytes, ifd: IFD, header: TIFFHeader,
     out_h = r1 - r0
     out_w = c1 - c0
 
-    _check_dimensions(out_w, out_h, samples, max_pixels)
+    _check_dimensions(out_w, out_h, samples, max_pixels, dtype=dtype)
 
     # StripByteCounts must have at least one entry per strip; a corrupt count
     # field can shrink it.  Detect the mismatch after the dimension safety
@@ -574,7 +574,7 @@ def _read_tiles(data: bytes, ifd: IFD, header: TIFFHeader,
     # what enforces the user's per-call memory cap. The source-read path
     # under ``read_vrt`` relies on that output check to honour a small
     # caller ``max_pixels`` against a normal-tile source.
-    _check_dimensions(tw, th, samples, MAX_PIXELS_DEFAULT)
+    _check_dimensions(tw, th, samples, MAX_PIXELS_DEFAULT, dtype=dtype)
 
     # Per-tile compressed-byte cap. Same env var as the
     # HTTP path. mmap slicing is bounded by the file size, but the slice
@@ -608,7 +608,7 @@ def _read_tiles(data: bytes, ifd: IFD, header: TIFFHeader,
     out_h = r1 - r0
     out_w = c1 - c0
 
-    _check_dimensions(out_w, out_h, samples, max_pixels)
+    _check_dimensions(out_w, out_h, samples, max_pixels, dtype=dtype)
 
     # Reject malformed TIFFs whose declared tile grid exceeds the number of
     # supplied TileOffsets entries. Silent skipping in the CPU loop below
